# product-catalog/src/main/resources/openapi/TMF-620-4-2.yaml
...
openapi: 3.0.3
info:
  title: Product Catalog Management TMF Adapter APIs
  version: 1.2.0
paths:
  /bsh-v1/productCatalogManagement/v5/productOffering:
    get:
      tags:
      - Tmf Product Offering Resource
      parameters:
      - name: accountId
        in: query
        description: Account Id
        required: true
        schema:
          type: string
      - name: id
        in: query
        description: id
        required: false
        schema:
          type: string
        example: 100
      - name: fields
        in: query
        description: Comma-separated properties to be provided in response
        required: false
        schema:
          type: string
        example: "id,name,status"
      - name: category
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          format: int32
          type: integer
      - name: name
        in: query
        schema:
          type: string
      - name: offset
        in: query
        schema:
          format: int32
          type: integer
      responses:
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "404":
          description: Product Offerings Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "200":
          description: Product Offering Found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductOffering'
      security:
      - Authorization: []
  /bsh-v1/productCatalogManagement/v5/productOffering/{id}:
    get:
      tags:
      - Tmf Product Offering Resource
      parameters:
      - name: accountId
        in: query
        description: Account Id
        required: true
        schema:
          type: string
      - name: id
        in: path
        description: id
        required: true
        schema:
          type: string
        example: 100
      - name: fields
        in: query
        description: Comma-separated properties to be provided in response
        required: false
        schema:
          type: string
        example: "id,name,status"
      responses:
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "404":
          description: Product Offering Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "200":
          description: Product Offering Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductOffering'
        "204":
          description: No Content
      security:
      - Authorization: []
  /bsh-v1/productCatalogManagement/v5/productOfferingPrice:
    get:
      tags:
      - Tmf Product Offering Price Resource
      parameters:
      - name: accountId
        in: query
        description: Account Id
        required: true
        schema:
          type: string
      - name: id
        in: query
        description: id
        required: false
        schema:
          type: string
        example: 100
      - name: fields
        in: query
        description: Comma-separated properties to be provided in response
        required: false
        schema:
          type: string
        example: "id,name,status"
      - name: limit
        in: query
        schema:
          format: int32
          type: integer
      - name: name
        in: query
        schema:
          type: string
      - name: offset
        in: query
        schema:
          format: int32
          type: integer
      responses:
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "404":
          description: Product Offering Price Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "200":
          description: Product Offering Price Found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductOfferingPrice'
      security:
      - Authorization: []
  /bsh-v1/productCatalogManagement/v5/productOfferingPrice/{id}:
    get:
      tags:
      - Tmf Product Offering Price Resource
      parameters:
      - name: accountId
        in: query
        description: Account Id
        required: true
        schema:
          type: string
      - name: productId
        in: query
        description: Product Id
        required: true
        schema:
          type: string
      - name: characteristic.name
        in: query
        description: Custom Field
        required: true
        schema:
          type: string
      - name: characteristic.value
        in: query
        description: Custom Field Value
        required: true
        schema:
          type: string
      - name: fields
        in: query
        description: Comma-separated properties to be provided in response
        required: false
        schema:
          type: string
        example: "id,name,status"
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "404":
          description: Product Offering Price Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "200":
          description: Product Offering Price Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductOfferingPrice'
        "204":
          description: No Content
      security:
      - Authorization: []
  /bsh-v1/productCatalogManagement/v5/productSpecification:
    get:
      tags:
      - Tmf Product Specification Resource
      parameters:
      - name: accountId
        in: query
        description: Account Id
        required: true
        schema:
          type: string
      - name: id
        in: query
        description: id
        required: false
        schema:
          type: string
        example: 100
      - name: fields
        in: query
        description: Comma-separated properties to be provided in response
        required: false
        schema:
          type: string
        example: "id,name,status"
      - name: brand
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          format: int32
          type: integer
      - name: name
        in: query
        schema:
          type: string
      - name: offset
        in: query
        schema:
          format: int32
          type: integer
      responses:
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "404":
          description: Product Specifications Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "200":
          description: Product Specification Found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductSpecification'
      security:
      - Authorization: []
  /bsh-v1/productCatalogManagement/v5/productSpecification/{id}:
    get:
      tags:
      - Tmf Product Specification Resource
      parameters:
      - name: accountId
        in: query
        description: Account Id
        required: true
        schema:
          type: string
      - name: fields
        in: query
        description: Comma-separated properties to be provided in response
        required: false
        schema:
          type: string
        example: "id,name,status"
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "404":
          description: Product Specification Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "200":
          description: Product Specification Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSpecification'
        "204":
          description: No Content
      security:
      - Authorization: []
  /bsh-v1/productCatalogManagement/v5/serviceSpecification/{vendor}/{provisionCategory}/{action}:
    get:
      tags:
      - Tmf Service Specification Resource
      parameters:
      - name: accountId
        in: query
        description: Account Id
        required: true
        schema:
          type: string
      - name: fields
        in: query
        description: Comma-separated properties to be provided in response
        required: false
        schema:
          type: string
        example: "id,name,status"
      - name: action
        in: path
        required: true
        schema:
          type: string
      - name: provisionCategory
        in: path
        required: true
        schema:
          type: string
      - name: vendor
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        "200":
          description: Product Offering Found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProvisioningSpecification'
      security:
      - Authorization: []
components:
  schemas:
...
    CustomError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        reason:
          type: string
...
    Entity:
      description: Base entity schema for use in TMForum Open-APIs. Property.
      type: object
      allOf:
      - $ref: '#/components/schemas/Extensible'
      properties:
        href:
          description: Hyperlink reference
          type: string
        id:
          description: unique identifier
          type: string
      discriminator:
        propertyName: '@type'
        mapping:
          Entity: '#/components/schemas/Entity'
...
    PlaceRef:
      description: Place reference.
      required:
      - id
      type: object
      allOf:
      - $ref: '#/components/schemas/Extensible'
      properties:
        href:
          description: The URI of the referred entity.
          type: string
        id:
          description: The identifier of the referred entity.
          type: string
        name:
          description: Name of the referred entity.
          type: string
        '@referredType':
          description: The actual type of the target instance when needed for disambiguation.
          type: string
      discriminator:
        propertyName: '@type'
        mapping:
          PlaceRef: '#/components/schemas/PlaceRef'
...
    ProductOffering:
      description: "ProductOffering represents entities that are orderable from the\
        \ provider of the catalog, this resource includes pricing information."
      type: object
      allOf:
      - $ref: '#/components/schemas/Entity'
      properties:
        description:
          description: Description of the productOffering
          type: string
        isBundle:
          description: "isBundle determines whether a productOffering represents a\
            \ single productOffering (false), or a bundle of productOfferings (true)."
          type: boolean
        isSellable:
          description: A flag indicating if this product offer can be sold stand-alone
            for sale or not. If this flag is false it indicates that the offer can
            only be sold within a bundle.
          type: boolean
        statusReason:
          description: A string providing a complementary information on the value
            of the lifecycle status attribute.
          type: string
        validFor:
          description: ""
          type: object
          allOf:
          - $ref: '#/components/schemas/TimePeriod'
        version:
          description: ProductOffering version
          type: string
        place:
          description: Place defines the places where the products are sold or delivered.
          type: array
          items:
            $ref: '#/components/schemas/PlaceRef'
        serviceLevelAgreement:
          description: ""
          type: object
          allOf:
          - $ref: '#/components/schemas/SLARef'
        channel:
          description: The channel defines the channel for selling product offerings.
          type: array
          items:
            $ref: '#/components/schemas/ChannelRef'
        serviceCandidate:
          description: ""
          type: object
          allOf:
          - $ref: '#/components/schemas/ServiceCandidateRef'
        category:
          description: "The category resource is used to group product offerings,\
            \ service and resource candidates in logical containers. Categories can\
            \ contain other categories and/or product offerings, resource or service\
            \ candidates."
          type: array
          items:
            $ref: '#/components/schemas/CategoryRef'
        resourceCandidate:
          description: ""
          type: object
          allOf:
          - $ref: '#/components/schemas/ResourceCandidateRef'
        productOfferingTerm:
          description: "A condition under which a ProductOffering is made available\
            \ to Customers. For instance, a productOffering can be offered with multiple\
            \ commitment periods."
          type: array
          items:
            $ref: '#/components/schemas/ProductOfferingTerm'
        productOfferingPrice:
          description: "An amount, usually of money, that is asked for or allowed\
            \ when a ProductOffering is bought, rented, or leased. The price is valid\
            \ for a defined period of time and may not represent the actual price\
            \ paid by a customer."
          type: array
          items:
            $ref: '#/components/schemas/ProductOfferingPriceRefOrValue'
        agreement:
          description: "An agreement represents a contract or arrangement, either\
            \ written or verbal and sometimes enforceable by law, such as a service\
            \ level agreement or a customer price agreement. An agreement involves\
            \ a number of other business entities, such as products, services, and\
            \ resources and/or their specifications."
          type: array
          items:
            $ref: '#/components/schemas/AgreementRef'
        bundledProductOffering:
          description: A type of ProductOffering that belongs to a grouping of ProductOfferings
            made available to the market. It inherits of all attributes of ProductOffering.
          type: array
          items:
            $ref: '#/components/schemas/BundledProductOffering'
        bundledGroupProductOffering:
          description: "A group of ProductOfferings that can be selected for instantiation,\
            \ e.g. between 2 and 7 from a list of 15 channel packs."
          type: array
          items:
            $ref: '#/components/schemas/BundledGroupProductOffering'
        attachment:
          description: "Complements the description of an element (for instance a\
            \ product) through video, pictures..."
          type: array
          items:
            $ref: '#/components/schemas/AttachmentRefOrValue'
        marketSegment:
          description: "provides references to the corresponding market segment as\
            \ target of product offerings. A market segment is grouping of Parties,\
            \ GeographicAreas, SalesChannels, and so forth."
          type: array
          items:
            $ref: '#/components/schemas/MarketSegmentRef'
        productOfferingRelationship:
          description: A relationship between this product offering and other product
            offerings.
          type: array
          items:
            $ref: '#/components/schemas/ProductOfferingRelationship'
        productOfferingCharacteristic:
          description: "A characteristic quality or distinctive feature of a ProductOffering.\
            \ The characteristic can take on a discrete value fixed at design (catalog\
            \ authoring) time, such as Mobile Plan Rank, and is not generally modifiable\
            \ at inventory level."
          type: array
          items:
            $ref: '#/components/schemas/CharacteristicSpecification'
        prodSpecCharValueUse:
          description: "A use of the ProductSpecificationCharacteristicValue by a\
            \ ProductOffering to which additional properties (attributes) apply or\
            \ override the properties of similar properties contained in ProductSpecificationCharacteristicValue.\
            \ It should be noted that characteristics which their value(s) addressed\
            \ by this object must exist in corresponding product specification. The\
            \ available characteristic values for a ProductSpecificationCharacteristic\
            \ in a Product specification can be modified at the ProductOffering level.\
            \ For example, a characteristic 'Color' might have values White, Blue,\
            \ Green, and Red. But, the list of values can be restricted to e.g. White\
            \ and Blue in an associated product offering. It should be noted that\
            \ the list of values in 'ProductSpecificationCharacteristicValueUse' is\
            \ a strict subset of the list of values as defined in the corresponding\
            \ product specification characteristics."
          type: array
          items:
            $ref: '#/components/schemas/ProductSpecificationCharacteristicValueUse'
        policy:
          description: The Policy resource represents a policy/rule applied to ProductOffering.
          type: array
          items:
            $ref: '#/components/schemas/PolicyRef'
        allowedAction:
          description: List of actions that can be executed (in context of a product
            order) on products instantiated from this offering
          type: array
          items:
            $ref: '#/components/schemas/AllowedProductAction'
        lastUpdate:
          format: date-time
          type: string
          example: 2024-04-17T16:30:25Z
        lifecycleStatus:
          description: Used to indicate the current lifecycle status
          type: string
        name:
          description: Name of the productOffering
          type: string
        productSpecification:
          description: ""
          type: object
          allOf:
          - $ref: '#/components/schemas/ProductSpecificationRef'
        externalIdentifier:
          description: "List of external identifieers for the offering, e.g. identifier\
            \ in source catalog"
          type: array
          items:
            $ref: '#/components/schemas/ExternalIdentifier'
      discriminator:
        propertyName: '@type'
        mapping:
          ProductOffering: '#/components/schemas/ProductOffering'
...
    ProductOfferingPrice:
      description: Price of ProductOffering
      type: object
      allOf:
      - $ref: '#/components/schemas/Entity'
      properties:
        description:
          description: Description of the productOfferingPrice
          type: string
        version:
          description: ProductOfferingPrice version
          type: string
        validFor:
          description: ""
          type: object
          allOf:
          - $ref: '#/components/schemas/TimePeriod'
        unitOfMeasure:
          description: ""
          type: object
          allOf:
          - $ref: '#/components/schemas/Quantity'
        recurringChargePeriodType:
          description: "The period to repeat the application of the price Could be\
            \ month, week..."
          type: string
        recurringChargePeriodLength:
          format: int32
          description: "the period of the recurring charge:  1, 2, ... .It sets to\
            \ zero if it is not applicable"
          type: integer
        isBundle:
          description: A flag indicating if this ProductOfferingPrice is composite
            (bundle) or not
          type: boolean
        price:
          description: ""
          type: object
          allOf:
          - $ref: '#/components/schemas/Money'
        percentage:
          format: float
          description: Percentage to apply if this Product Offering Price is an Alteration
            (such as a Discount)
          type: number
        bundledPopRelationship:
          description: this object represents a bundle relationship from a bundle
            product offering price (parent) to a simple product offering price (child).
            A simple product offering price may participate in more than one bundle
            relationship.
          type: array
          items:
            $ref: '#/components/schemas/BundledProductOfferingPriceRelationship'
        popRelationship:
          description: "Product Offering Prices related to this Product Offering Price,\
            \ for example a price alteration such as allowance or discount"
          type: array
          items:
            $ref: '#/components/schemas/ProductOfferingPriceRelationship'
        prodSpecCharValueUse:
          description: A use of the ProductSpecificationCharacteristicValue by a ProductOfferingPrice
            to which additional properties (attributes) apply or override the properties
            of similar properties contained in ProductSpecificationCharacteristicValue.
            It should be noted that characteristics which their value(s) addressed
            by this object must exist in corresponding product specification. The
            available characteristic values for a ProductSpecificationCharacteristic
            in a Product specification can be modified at the ProductOffering and
            ProcuctOfferingPrice level. The list of values in ProductSpecificationCharacteristicValueUse
            is a strict subset of the list of values as defined in the corresponding
            product specification characteristics.
          type: array
          items:
            $ref: '#/components/schemas/ProductSpecificationCharacteristicValueUse'
        productOfferingTerm:
          description: "A list of conditions under which a ProductOfferingPrice is\
            \ made available to Customers. For instance, a Product Offering Price\
            \ can be offered with multiple commitment periods."
          type: array
          items:
            $ref: '#/components/schemas/ProductOfferingTerm'
        place:
          description: Place defines the places where the products are sold or delivered.
          type: array
          items:
            $ref: '#/components/schemas/PlaceRef'
        policy:
          description: The Policy resource represents a policy/rule applied to ProductOfferingPrice.
          type: array
          items:
            $ref: '#/components/schemas/PolicyRef'
        pricingLogicAlgorithm:
          description: The PricingLogicAlgorithm entity represents an instantiation
            of an interface specification to external rating function (without a modeled
            behavior in SID). Some of the parameters of the interface definition may
            be already set (such as price per unit) and some may be gathered during
            the rating process from the event (such as call duration) or from ProductCharacteristicValues
            (such as assigned bandwidth).
          type: array
          items:
            $ref: '#/components/schemas/PricingLogicAlgorithm'
        tax:
          description: An amount of money levied on the price of a Product by a legislative
            body.
          type: array
          items:
            $ref: '#/components/schemas/TaxItem'
        name:
          description: Name of the productOfferingPrice
          type: string
        priceType:
          description: "A category that describes the price, such as recurring, discount,\
            \ allowance, penalty, and so forth."
          type: string
        lastUpdate:
          format: date-time
          type: string
          example: 2024-04-17T16:30:25Z
        lifecycleStatus:
          description: the lifecycle status of this ProductOfferingPrice
          type: string
        externalIdentifier:
          description: "List of external identifieers for the offering price, e.g.\
            \ identifier in source catalog"
          type: array
          items:
            $ref: '#/components/schemas/ExternalIdentifier'
        bshPhaseType:
          type: string
        bshPhases:
          type: array
          items:
            $ref: '#/components/schemas/Phase'
        productName:
          type: string
        pricebookPrice:
          $ref: '#/components/schemas/Money'
        pricebookMasterPrice:
          $ref: '#/components/schemas/Money'
        accountCost:
          $ref: '#/components/schemas/Money'
      discriminator:
        propertyName: '@type'
        mapping:
          ProductOfferingPrice: '#/components/schemas/ProductOfferingPrice'
...
    ProvisioningSpecification:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        lastUpdate:
          format: date-time
          type: string
          example: 2024-04-17T16:30:25Z
        lifecycleStatus:
          type: string
        version:
          type: string
        provisioningSpecCharacteristic:
          type: array
          items:
            $ref: '#/components/schemas/ProvisioningSpecCharacteristic'
    Quantity:
      description: An amount in a given unit
      type: object
      properties:
        amount:
          format: float
          description: Numeric value in a given unit
          type: number
        units:
          description: Unit
          type: string
...
    SLARef:
      description: "ServiceLevelAgreement reference: A service level agreement (SLA)\
        \ is a type of agreement that represents a formal negotiated agreement between\
        \ two parties designed to create a common understanding about products, services,\
        \ priorities, responsibilities, and so forth. The SLA is a set of appropriate\
        \ procedures and targets formally or informally agreed between parties in\
        \ order to achieve and maintain specified Quality of Service."
      required:
      - id
      type: object
      allOf:
      - $ref: '#/components/schemas/Extensible'
      properties:
        href:
          description: Hyperlink reference
          type: string
        id:
          description: unique identifier
          type: string
        name:
          description: Name of the referred entity.
          type: string
        '@referredType':
          description: The actual type of the target instance when needed for disambiguation.
          type: string
      discriminator:
        propertyName: '@type'
        mapping:
          SLARef: '#/components/schemas/SLARef'
...
    TimePeriod:
      description: "A period of time, either as a deadline (endDateTime only) a startDateTime\
        \ only, or both"
      type: object
      properties:
        startDateTime:
          description: "Start of the time period, using IETC-RFC-3339 format"
          type: string
          allOf:
          - $ref: '#/components/schemas/Date'
          example: 1985-04-12T23:20:50.520Z
        endDateTime:
          description: "End of the time period, using IETC-RFC-3339 format"
          type: string
          allOf:
          - $ref: '#/components/schemas/Date'
          example: 1985-04-12T23:20:50.520Z
  securitySchemes:
    Authorization:
      type: http
      scheme: Basic